Added settings menu for the players
The menu is almost fully functional (you can try it with the field and
background color change) and provides an idea of just how it should work
(it is not much styled yet). Values for some of the options still need
to be discussed and implemented soon. All the values it stored are
easily accessible through the meny methods in settings.js . For further
information about the menu and suggestions contact YoTsenkov(Dancho) ;D